FIX - Issue #32: Fixed the issue with the tests by changing the positions of the files
